Added predicates
