+++ python/CMakeLists.txt
@@ -25,6 +25,12 @@ endif()
 
 set(CSVPY_SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/csvpy.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/csvpy_core.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/csvpy_document.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/csvpy_format.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/csvpy_numpy.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/csvpy_predicate.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/csvpy_reader.cpp
 )
 
 nanobind_add_module(csvpy NOMINSIZE ${CSVPY_SOURCES})
