+++ tests/shared/generated_file.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdlib>
 #include <fstream>
 #include <string>
 #include <utility>
@@ -17,7 +18,7 @@ namespace csv_test {
     class GeneratedFile {
     public:
         explicit GeneratedFile(std::string filename)
-            : guard_(std::move(filename)) {}
+            : guard_(temp_filename(std::move(filename))) {}
 
         template<typename Generator>
         const std::string& path(Generator&& generator) {
@@ -31,6 +32,52 @@ namespace csv_test {
         }
 
     private:
+        static std::string env_value(const char* name) {
+#if defined(_MSC_VER)
+            char* value = nullptr;
+            size_t size = 0;
+            if (_dupenv_s(&value, &size, name) != 0 || value == nullptr) {
+                return std::string();
+            }
+
+            std::string result(value);
+            std::free(value);
+            return result;
+#else
+            const char* value = std::getenv(name);
+            return value ? std::string(value) : std::string();
+#endif
+        }
+
+        static std::string temp_filename(std::string filename) {
+            if (filename.find('/') != std::string::npos || filename.find('\\') != std::string::npos) {
+                return filename;
+            }
+
+            std::string temp_dir = env_value("TMPDIR");
+            if (temp_dir.empty()) {
+                temp_dir = env_value("TEMP");
+            }
+            if (temp_dir.empty()) {
+                temp_dir = env_value("TMP");
+            }
+            if (temp_dir.empty()) {
+                return filename;
+            }
+
+            std::string path(std::move(temp_dir));
+            const char last = path.empty() ? '\0' : path[path.size() - 1];
+            if (last != '/' && last != '\\') {
+#if defined(_WIN32)
+                path.push_back('\\');
+#else
+                path.push_back('/');
+#endif
+            }
+            path += filename;
+            return path;
+        }
+
         FileGuard guard_;
         bool generated_ = false;
     };
